# gdal raster reproject: fix performance issue when outputing to COG