--- a/apps/gdalalg_raster_aspect.h
+++ b/apps/gdalalg_raster_aspect.h
@@ -34,7 +34,7 @@ class GDALRasterAspectAlgorithm /* non final */
     explicit GDALRasterAspectAlgorithm(bool standaloneStep = false);
 
   private:
-    bool RunStep(GDALProgressFunc pfnProgress, void *pProgressData) override;
+    bool RunStep(GDALRasterPipelineStepRunContext &ctxt) override;
 
     int m_band = 1;
     std::string m_convention = "azimuth";