--- a/apps/gdalalg_raster_color_map.h
+++ b/apps/gdalalg_raster_color_map.h
@@ -37,7 +37,7 @@ class GDALRasterColorMapAlgorithm /* non final */
     explicit GDALRasterColorMapAlgorithm(bool standaloneStep = false);
 
   private:
-    bool RunStep(GDALProgressFunc pfnProgress, void *pProgressData) override;
+    bool RunStep(GDALRasterPipelineStepRunContext &ctxt) override;
 
     int m_band = 1;
     std::string m_colorMap{};