--- a/apps/gdalalg_raster_fill_nodata.h
+++ b/apps/gdalalg_raster_fill_nodata.h
@@ -35,7 +35,7 @@ class GDALRasterFillNodataAlgorithm /* non final */
         bool standaloneStep = false) noexcept;
 
   private:
-    bool RunStep(GDALProgressFunc pfnProgress, void *pProgressData) override;
+    bool RunStep(GDALRasterPipelineStepRunContext &ctxt) override;
 
     // The maximum distance (in pixels) that the algorithm will search out for values to interpolate. The default is 100 pixels.
     int m_maxDistance = 100;