--- a/apps/gdalalg_raster_hillshade.h
+++ b/apps/gdalalg_raster_hillshade.h
@@ -35,7 +35,7 @@ class GDALRasterHillshadeAlgorithm /* non final */
     explicit GDALRasterHillshadeAlgorithm(bool standaloneStep = false);
 
   private:
-    bool RunStep(GDALProgressFunc pfnProgress, void *pProgressData) override;
+    bool RunStep(GDALRasterPipelineStepRunContext &ctxt) override;
 
     int m_band = 1;
     double m_zfactor = 1;