--- a/apps/gdalalg_raster_pipeline.cpp
+++ b/apps/gdalalg_raster_pipeline.cpp
@@ -172,42 +172,47 @@ bool GDALRasterPipelineStepAlgorithm::RunImpl(GDALProgressFunc pfnProgress,
         CPLAssert(!m_executionForStreamOutput || bIsStreaming);
 
         bool ret = false;
-        if (readAlg.Run())
+        if (!m_outputVRTCompatible &&
+            (EQUAL(m_format.c_str(), "VRT") ||
+             (m_format.empty() &&
+              EQUAL(CPLGetExtensionSafe(m_outputDataset.GetName().c_str())
+                        .c_str(),
+                    "VRT"))))
+        {
+            ReportError(CE_Failure, CPLE_NotSupported,
+                        "VRT output is not supported. Consider using the "
+                        "GDALG driver instead (files with .gdalg.json "
+                        "extension)");
+        }
+        else if (readAlg.Run())
         {
             m_inputDataset.Set(readAlg.m_outputDataset.GetDatasetRef());
             m_outputDataset.Set(nullptr);
 
             std::unique_ptr<void, decltype(&GDALDestroyScaledProgress)>
                 pScaledData(nullptr, GDALDestroyScaledProgress);
-            if (pfnProgress && !IsNativelyStreamingCompatible())
+
+            const bool bCanHandleNextStep =
+                !bIsStreaming && CanHandleNextStep(&writeAlg);
+
+            if (pfnProgress &&
+                (bCanHandleNextStep || !IsNativelyStreamingCompatible()))
             {
                 pScaledData.reset(GDALCreateScaledProgress(
-                    0.0, bIsStreaming ? 1.0 : 0.5, pfnProgress, pProgressData));
+                    0.0, bIsStreaming || bCanHandleNextStep ? 1.0 : 0.5,
+                    pfnProgress, pProgressData));
             }
 
             GDALRasterPipelineStepRunContext stepCtxt;
             stepCtxt.m_pfnProgress = pScaledData ? GDALScaledProgress : nullptr;
             stepCtxt.m_pProgressData = pScaledData.get();
+            if (bCanHandleNextStep)
+                stepCtxt.m_poNextStep = &writeAlg;
             if (RunPreStepPipelineValidations() && RunStep(stepCtxt))
             {
-                if (bIsStreaming)
+                if (bIsStreaming || bCanHandleNextStep)
                 {
                     ret = true;
-                }
-                else if (!m_outputVRTCompatible &&
-                         (EQUAL(m_format.c_str(), "VRT") ||
-                          (m_format.empty() &&
-                           EQUAL(CPLGetExtensionSafe(
-                                     writeAlg.m_outputDataset.GetName().c_str())
-                                     .c_str(),
-                                 "VRT"))))
-                {
-                    ReportError(
-                        CE_Failure, CPLE_NotSupported,
-                        "VRT output is not supported. Consider using the "
-                        "GDALG driver instead (files with .gdalg.json "
-                        "extension)");
-                    ret = false;
                 }
                 else
                 {