--- a/apps/gdalalg_raster_proximity.h
+++ b/apps/gdalalg_raster_proximity.h
@@ -34,7 +34,7 @@ class GDALRasterProximityAlgorithm /* non final */
     explicit GDALRasterProximityAlgorithm(bool standaloneStep = false);
 
   private:
-    bool RunStep(GDALProgressFunc pfnProgress, void *pProgressData) override;
+    bool RunStep(GDALRasterPipelineStepRunContext &ctxt) override;
 
     double m_noDataValue = 0.0;
     int m_inputBand = 1;