--- a/apps/gdalalg_raster_read.cpp
+++ b/apps/gdalalg_raster_read.cpp
@@ -33,7 +33,7 @@ GDALRasterReadAlgorithm::GDALRasterReadAlgorithm()
 /*                  GDALRasterReadAlgorithm::RunStep()                  */
 /************************************************************************/
 
-bool GDALRasterReadAlgorithm::RunStep(GDALProgressFunc, void *)
+bool GDALRasterReadAlgorithm::RunStep(GDALRasterPipelineStepRunContext &)
 {
     CPLAssert(m_inputDataset.GetDatasetRef());
     CPLAssert(m_outputDataset.GetName().empty());