… apps/gdalalg_raster_reproject.h
… apps/gdalalg_raster_resize.cpp
… apps/gdalalg_raster_resize.h
… apps/gdalalg_raster_rgb_to_palette.cpp
… apps/gdalalg_raster_rgb_to_palette.h
… apps/gdalalg_raster_roughness.cpp
… apps/gdalalg_raster_roughness.h
… apps/gdalalg_raster_scale.cpp
… apps/gdalalg_raster_scale.h
… apps/gdalalg_raster_select.cpp
… apps/gdalalg_raster_select.h
… apps/gdalalg_raster_set_type.cpp
… apps/gdalalg_raster_set_type.h
… apps/gdalalg_raster_sieve.cpp
… apps/gdalalg_raster_sieve.h
… apps/gdalalg_raster_slope.cpp
… apps/gdalalg_raster_slope.h
… apps/gdalalg_raster_stack.cpp
… apps/gdalalg_raster_tpi.cpp
… apps/gdalalg_raster_tpi.h
… apps/gdalalg_raster_tri.cpp
… apps/gdalalg_raster_tri.h
… apps/gdalalg_raster_unscale.cpp
… apps/gdalalg_raster_unscale.h
… apps/gdalalg_raster_viewshed.cpp
… apps/gdalalg_raster_viewshed.h
… apps/gdalalg_raster_write.cpp
… apps/gdalalg_raster_write.h
… apps/gdalalg_vector_clip.cpp
… apps/gdalalg_vector_clip.h