We recover performance by allowing a step to do the job of the following step, if it declares to be able to do so. This enables us to replace warp to VRT + translate  with warp to final format, which is very beneficial for COG output given that there is a co-operation between the COG driver and the warping code.

With this PR:
```
$ time gdal raster pipeline read byte.tif ! reproject --size 10000,10000 -r cubic --t_srs EPSG:4326 -j 0 ! write out.tif --overwrite --of=cog --co compress=zstd

real    0m5,748s
user    0m5,595s
sys     0m0,152s

$ time gdalwarp byte.tif out.tif -ts 10000 10000 -r cubic -wo NUM_THREADS=0 -co NUM_THREADS=0 -of COG -co COMPRESS=ZSTD -overwrite

real    0m5,280s
user    0m5,110s
sys     0m0,168s

$ time gdal raster pipeline read byte.tif ! reproject --size 10000,10000 -r cubic --t_srs EPSG:4326 -j 8 ! write out.tif --overwrite --of=cog --co compress=zstd

real    0m2,599s
user    0m11,262s
sys     0m0,527s

$ time gdal raster reproject --size 10000,10000 -r cubic --t_srs EPSG:4326 -j 8 --overwrite --of=cog --co compress=zstd byte.tif out.tif

real	0m2,594s
user	0m11,633s
sys	0m0,464s

$ time gdalwarp byte.tif out.tif -ts 10000 10000 -r cubic -wo NUM_THREADS=8 -co NUM_THREADS=8 -of COG -co COMPRESS=ZSTD -overwrite
Creating output file that is 10000P x 10000L.

real    0m2,258s
user    0m10,186s
sys     0m0,344s
```

Compared to master + https://github.com/OSGeo/gdal/pull/12383:
```
$ time gdal raster pipeline read byte.tif ! reproject --size 10000,10000 -r cubic --t_srs EPSG:4326 -j 0 ! write out.tif --overwrite --of=cog --co compress=zstd

real    0m8,631s
user    0m8,558s
sys     0m0,072s

$ time gdal raster pipeline read byte.tif ! reproject --size 10000,10000 -r cubic --t_srs EPSG:4326 -j 8 ! write out.tif --overwrite --of=cog --co compress=zstd --co num_threads=8

real    0m6,256s
user    0m17,601s
sys     0m0,388s
```